requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
qdrant-client==1.15.0
pandas==2.3.1
sentence-transformers==5.0.0
python-dotenv==1.1.1
requests==2.32.4
redis==6.2.0
prometheus-fastapi-instrumentator==7.1.0
pytest==8.4.1
httpx==0.28.1